Support for AggroInfo long numbers.

+++ L2J_Mobius_04.0_GrandCrusade/java/org/l2jmobius/gameserver/ai/AttackableAI.java
@@ -331,7 +331,7 @@ public class AttackableAI extends CreatureAI
 							|| (Config.FAKE_PLAYER_AGGRO_MONSTERS && t.isMonster() && !t.isFakePlayer()) //
 							|| (Config.FAKE_PLAYER_AGGRO_PLAYERS && t.isPlayer()))
 						{
-							final int hating = npc.getHating(t);
+							final long hating = npc.getHating(t);
 							final double distance = npc.calculateDistance2D(t);
 							if ((hating == 0) && (closestDistance > distance))
 							{
@@ -392,7 +392,7 @@ public class AttackableAI extends CreatureAI
 						{
 							if (!npc.isFakePlayer() || (npc.isFakePlayer() && Config.FAKE_PLAYER_AGGRO_FPC))
 							{
-								final int hating = npc.getHating(t);
+								final long hating = npc.getHating(t);
 								if (hating == 0)
 								{
 									npc.addDamageHate(t, 0, 0);
@@ -411,7 +411,7 @@ public class AttackableAI extends CreatureAI
 							}
 							
 							// Get the hate level of the Attackable against this Creature target contained in _aggroList
-							final int hating = npc.getHating(t);
+							final long hating = npc.getHating(t);
 							
 							// Add the attacker to the Attackable _aggroList with 0 damage and 1 hate
 							if (hating == 0)
@@ -442,7 +442,7 @@ public class AttackableAI extends CreatureAI
 			if ((hated != null) && !npc.isCoreAIDisabled())
 			{
 				// Get the hate level of the Attackable against this Creature target contained in _aggroList
-				final int aggro = npc.getHating(hated);
+				final long aggro = npc.getHating(hated);
 				if ((aggro + _globalAggro) > 0)
 				{
 					// Set the Creature movement type to run and send Server->Client packet ChangeMoveType to all others Player
@@ -1239,7 +1239,7 @@ public class AttackableAI extends CreatureAI
 			}
 		}
 		
-		int searchValue = Integer.MIN_VALUE;
+		long searchValue = Long.MIN_VALUE;
 		Creature creature = null;
 		for (AggroInfo aggro : npc.getAggroList().values())
 		{
