+++ L2J_Mobius_05.5_EtinasFate/java/org/l2jmobius/gameserver/model/actor/instance/FriendlyNpc.java
@@ -79,7 +79,7 @@ public class FriendlyNpc extends Attackable
 	}
 	
 	@Override
-	public void addDamageHate(Creature attacker, int damage, int aggro)
+	public void addDamageHate(Creature attacker, long damage, long aggro)
 	{
 		if (!attacker.isPlayable() && !(attacker instanceof FriendlyNpc))
 		{
