+++ L2J_Mobius_CT_0_Interlude/java/org/l2jmobius/gameserver/model/actor/instance/Defender.java
@@ -164,7 +164,7 @@ public class Defender extends Attackable
 	}
 	
 	@Override
-	public void addDamageHate(Creature attacker, int damage, int aggro)
+	public void addDamageHate(Creature attacker, long damage, long aggro)
 	{
 		if (attacker == null)
 		{
