+++ L2J_Mobius_CT_2.4_Epilogue/java/org/l2jmobius/gameserver/ai/SiegeGuardAI.java
@@ -275,7 +275,7 @@ public class SiegeGuardAI extends CreatureAI implements Runnable
 			if (hated != null)
 			{
 				// Get the hate level of the Attackable against this Creature target contained in _aggroList
-				final int aggro = npc.getHating(hated);
+				final long aggro = npc.getHating(hated);
 				if ((aggro + _globalAggro) > 0)
 				{
 					// Set the Creature movement type to run and send Server->Client packet ChangeMoveType to all others Player
@@ -493,7 +493,7 @@ public class SiegeGuardAI extends CreatureAI implements Runnable
 						{
 							break;
 						}
-
+						
 						boolean useSkillSelf = true;
 						if (sk.isContinuous() && !sk.isDebuff() && _actor.isAffectedBySkill(sk.getId()))
 						{
