+++ L2J_Mobius_CT_2.6_HighFive/java/org/l2jmobius/gameserver/model/actor/instance/FortCommander.java
@@ -67,7 +67,7 @@ public class FortCommander extends Defender
 	}
 	
 	@Override
-	public void addDamageHate(Creature attacker, int damage, int aggro)
+	public void addDamageHate(Creature attacker, long damage, long aggro)
 	{
 		if (attacker == null)
 		{
