+++ L2J_Mobius_Classic_2.9.5_Saviors/java/org/l2jmobius/gameserver/model/actor/Attackable.java
@@ -737,7 +737,7 @@ public class Attackable extends Npc
 			return null;
 		}
 		
-		int damage = 0;
+		long damage = 0;
 		Creature damageDealer = null;
 		for (AggroInfo info : _aggroList.values())
 		{
@@ -805,7 +805,7 @@ public class Attackable extends Npc
 	 * @param damage The number of damages given by the attacker Creature
 	 * @param aggroValue The hate (=damage) given by the attacker Creature
 	 */
-	public void addDamageHate(Creature creature, int damage, int aggroValue)
+	public void addDamageHate(Creature creature, long damage, long aggroValue)
 	{
 		Creature attacker = creature;
 		if ((attacker == null) || (attacker == this))
@@ -834,7 +834,7 @@ public class Attackable extends Npc
 		// traps does not cause aggro
 		// making this hack because not possible to determine if damage made by trap
 		// so just check for triggered trap here
-		int aggro = aggroValue;
+		long aggro = aggroValue;
 		if ((targetPlayer == null) || (targetPlayer.getTrap() == null) || !targetPlayer.getTrap().isTriggered())
 		{
 			ai.addHate(aggro);
@@ -869,7 +869,7 @@ public class Attackable extends Npc
 		}
 	}
 	
-	public void reduceHate(Creature target, int amount)
+	public void reduceHate(Creature target, long amount)
 	{
 		if (target == null) // whole aggrolist
 		{
@@ -946,7 +946,7 @@ public class Attackable extends Npc
 		}
 		
 		Creature mostHated = null;
-		int maxHate = 0;
+		long maxHate = 0;
 		
 		// While Interacting over This Map Removing Object is Not Allowed
 		// Go through the aggroList of the Attackable
@@ -979,7 +979,7 @@ public class Attackable extends Npc
 		
 		Creature mostHated = null;
 		Creature secondMostHated = null;
-		int maxHate = 0;
+		long maxHate = 0;
 		final List<Creature> result = new ArrayList<>();
 		
 		// While iterating over this map removing objects is not allowed
@@ -1029,7 +1029,7 @@ public class Attackable extends Npc
 	 * @param target The Creature whose hate level must be returned
 	 * @return the hate level of the Attackable against this Creature contained in _aggroList.
 	 */
-	public int getHating(Creature target)
+	public long getHating(Creature target)
 	{
 		if (_aggroList.isEmpty() || (target == null))
 		{
