+++ L2J_Mobius_Classic_Interlude/java/org/l2jmobius/gameserver/model/actor/Attackable.java
@@ -721,7 +721,7 @@ public class Attackable extends Npc
 			return null;
 		}
 		
-		int damage = 0;
+		long damage = 0;
 		Creature damageDealer = null;
 		for (AggroInfo info : _aggroList.values())
 		{
@@ -789,7 +789,7 @@ public class Attackable extends Npc
 	 * @param damage The number of damages given by the attacker Creature
 	 * @param aggroValue The hate (=damage) given by the attacker Creature
 	 */
-	public void addDamageHate(Creature creature, int damage, int aggroValue)
+	public void addDamageHate(Creature creature, long damage, long aggroValue)
 	{
 		Creature attacker = creature;
 		if ((attacker == null) || (attacker == this))
@@ -818,7 +818,7 @@ public class Attackable extends Npc
 		// traps does not cause aggro
 		// making this hack because not possible to determine if damage made by trap
 		// so just check for triggered trap here
-		int aggro = aggroValue;
+		long aggro = aggroValue;
 		if ((targetPlayer == null) || (targetPlayer.getTrap() == null) || !targetPlayer.getTrap().isTriggered())
 		{
 			ai.addHate(aggro);
@@ -853,7 +853,7 @@ public class Attackable extends Npc
 		}
 	}
 	
-	public void reduceHate(Creature target, int amount)
+	public void reduceHate(Creature target, long amount)
 	{
 		if (target == null) // whole aggrolist
 		{
@@ -930,7 +930,7 @@ public class Attackable extends Npc
 		}
 		
 		Creature mostHated = null;
-		int maxHate = 0;
+		long maxHate = 0;
 		
 		// While Interacting over This Map Removing Object is Not Allowed
 		// Go through the aggroList of the Attackable
@@ -963,7 +963,7 @@ public class Attackable extends Npc
 		
 		Creature mostHated = null;
 		Creature secondMostHated = null;
-		int maxHate = 0;
+		long maxHate = 0;
 		final List<Creature> result = new ArrayList<>();
 		
 		// While iterating over this map removing objects is not allowed
@@ -1013,7 +1013,7 @@ public class Attackable extends Npc
 	 * @param target The Creature whose hate level must be returned
 	 * @return the hate level of the Attackable against this Creature contained in _aggroList.
 	 */
-	public int getHating(Creature target)
+	public long getHating(Creature target)
 	{
 		if (_aggroList.isEmpty() || (target == null))
 		{
